small refactor of test utils
